Extract class to other file
(cherry picked from commit 78425101796f85032bdeb9c5a5554256d5674ce2)
